skipping the sde tests when output file pre exist

+++ scripts/batch_jobs/run_sde_knl.sh
@@ -7,16 +7,17 @@
 CAFFE_ROOT=/project/projectdirs/mpccc/tmalas/intelcaffe/src/
 MODEL=train_val.prototxt
 SDE_PARSER="/global/homes/t/tmalas/ai_scripts/parse-sde.sh" # available here: http://portal.nersc.gov/project/m888/examples/stream-ai-example_160518.tgz
+module unload intel
 
 # CORI/GERTY
-#module switch intel intel/16.0.3.210.test
+#module load intel/16.0.3.210.test
 #export OMP_PLACES=cores"(68)"
 #export OMP_NUM_THREADS=68
 #execdir=/project/projectdirs/mpccc/tmalas/intelcaffe/install_cori-knl/bin
 #exe_pre="srun -n 1 -c 272 --cpu_bind=cores "
 
 # CARL
-module switch intel intel/17.0.0.098
+module load intel/17.0.0.098
 export OMP_PLACES=cores"(64)"
 export OMP_NUM_THREADS=64
 execdir=/project/projectdirs/mpccc/tmalas/intelcaffe/install_carl/bin
@@ -34,9 +35,11 @@ export layers_count="$(cat ${MODEL} | grep layer | wc -l)"
 
 # run once to get the layers time
 export out_file=${results_dir}/knl_time.out
-exe=${exe_pre}"${execdir}/caffe time -model=${MODEL} -iterations=100"
-echo $exe  | tee ${out_file}
-${exe} 2>&1 | tee -a ${out_file}
+if [ ! -f ${out_file} ]; then
+  exe=${exe_pre}"${execdir}/caffe time -model=${MODEL} -iterations=100"
+  echo $exe  | tee ${out_file}
+  ${exe} 2>&1 | tee -a ${out_file}
+fi
  
 export OMP_NUM_THREADS=16
 for direction in 1 0
@@ -44,10 +47,12 @@ do
   for layer in $(seq 0 `expr $layers_count - 1`)
   do
     export out_file=${results_dir}/knl_sde_layer${layer}_dir${direction}.out
-    exe=${exe_pre}"sde64 -knl -d -iform 1 -omix ${out_file}.mix -global_region -start_ssc_mark 111:repeat -stop_ssc_mark 222:repeat  -- ${execdir}/caffe time -model=${MODEL} -iterations=1 -prof_layer=${layer} -prof_forward_direction=${direction}"
-    echo $exe  | tee ${out_file}
-    ${exe} 2>&1 | tee -a ${out_file}
-    ${SDE_PARSER} ${out_file}.mix | tee -a ${out_file}
+    if [ ! -f ${out_file} ]; then
+      exe=${exe_pre}"sde64 -knl -d -iform 1 -omix ${out_file}.mix -global_region -start_ssc_mark 111:repeat -stop_ssc_mark 222:repeat  -- ${execdir}/caffe time -model=${MODEL} -iterations=1 -prof_layer=${layer} -prof_forward_direction=${direction}"
+      echo $exe   | tee ${out_file}
+      ${exe} 2>&1 | tee -a ${out_file}
+      ${SDE_PARSER} ${out_file}.mix | tee -a ${out_file}
+    fi
   done
 done
